server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
ariadne==0.24.0
databases==0.9.0
SQLAlchemy==2.0.37
pyjwt==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1